adding middleware for retrieving membership
