+++ src/controllers/membership/index.ts
@@ -1,9 +1,10 @@
 import { default as create } from './create';
 import { default as getAll } from './getAll';
-import { default as getOne } from './getOne';
+import { default as getOne, getMembershipMiddleware } from './getOne';
 
 export default {
   create,
   getAll,
   getOne,
+  getMembershipMiddleware,
 };
