+++ src/server/server.ts
@@ -10,7 +10,7 @@ import {
   configureRoutes,
   initializeDatabaseConnection,
 } from './utils';
-import { User, Project } from 'src/models';
+import { User, Project, Membership } from 'src/models';
 
 // Extend the types availble on the Express request/response objects.
 declare global {
@@ -19,6 +19,7 @@ declare global {
     interface Request {
       user: User;
       project: Project;
+      membership: Membership;
     }
     interface Response {
       /* eslint-disable-next-line @typescript-eslint/no-explicit-any */
